handtracking/requirements.txt
pygame==1.9.4
open3d==0.8.0.0
tensorflow_gpu==1.14.0
transforms3d==0.3.1
keyboard==0.13.4
opencv_python==3.4.3.18
numpy>=1.18.1
six
wcwidth
pillow
matplotlib
mediapipe
pyzmq
